// contracts/typechain-types/ViewProfile.ts
export declare namespace ViewProfile {
  export type TopicViewStruct = {
    sentiment: BigNumberish;
    updatedAt: BigNumberish;
  };

  export type TopicViewStructOutput = [sentiment: bigint, updatedAt: bigint] & {
    sentiment: bigint;
    updatedAt: bigint;
  };
}

export interface ViewProfileInterface extends Interface {
  getFunction(
    nameOrSignature: "getTopicView" | "hasTopicView" | "setTopicView"
  ): FunctionFragment;

  getEvent(nameOrSignatureOrTopic: "TopicViewUpdated"): EventFragment;

  encodeFunctionData(
    functionFragment: "getTopicView",
    values: [AddressLike, BytesLike]
  ): string;
  encodeFunctionData(
    functionFragment: "hasTopicView",
    values: [AddressLike, BytesLike]
  ): string;
  encodeFunctionData(
    functionFragment: "setTopicView",
    values: [BytesLike, BigNumberish]
  ): string;

  decodeFunctionResult(
    functionFragment: "getTopicView",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "hasTopicView",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "setTopicView",
    data: BytesLike
  ): Result;
}

export namespace TopicViewUpdatedEvent {
  export type InputTuple = [
    user: AddressLike,
    topicId: BytesLike,
    sentiment: BigNumberish,
    updatedAt: BigNumberish
  ];
  export type OutputTuple = [
    user: string,
    topicId: string,
    sentiment: bigint,
    updatedAt: bigint
  ];
  export interface OutputObject {
    user: string;
    topicId: string;
    sentiment: bigint;
    updatedAt: bigint;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export interface ViewProfile extends BaseContract {
  connect(runner?: ContractRunner | null): ViewProfile;
  waitForDeployment(): Promise<this>;

  interface: ViewProfileInterface;

  queryFilter<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;
  queryFilter<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;

  on<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  on<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  once<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  once<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  listeners<TCEvent extends TypedContractEvent>(
    event: TCEvent
  ): Promise<Array<TypedListener<TCEvent>>>;
  listeners(eventName?: string): Promise<Array<Listener>>;
  removeAllListeners<TCEvent extends TypedContractEvent>(
    event?: TCEvent
  ): Promise<this>;

  getTopicView: TypedContractMethod<
    [user: AddressLike, topicId: BytesLike],
    [ViewProfile.TopicViewStructOutput],
    "view"
  >;

  hasTopicView: TypedContractMethod<
    [user: AddressLike, topicId: BytesLike],
    [boolean],
    "view"
  >;

  setTopicView: TypedContractMethod<
    [topicId: BytesLike, sentiment: BigNumberish],
    [void],
    "nonpayable"
  >;

  getFunction<T extends ContractMethod = ContractMethod>(
    key: string | FunctionFragment
  ): T;

  getFunction(
    nameOrSignature: "getTopicView"
  ): TypedContractMethod<
    [user: AddressLike, topicId: BytesLike],
    [ViewProfile.TopicViewStructOutput],
    "view"
  >;
  getFunction(
    nameOrSignature: "hasTopicView"
  ): TypedContractMethod<
    [user: AddressLike, topicId: BytesLike],
    [boolean],
    "view"
  >;
  getFunction(
    nameOrSignature: "setTopicView"
  ): TypedContractMethod<
    [topicId: BytesLike, sentiment: BigNumberish],
    [void],
    "nonpayable"
  >;

  getEvent(
    key: "TopicViewUpdated"
  ): TypedContractEvent<
    TopicViewUpdatedEvent.InputTuple,
    TopicViewUpdatedEvent.OutputTuple,
    TopicViewUpdatedEvent.OutputObject
  >;

  filters: {
    "TopicViewUpdated(address,bytes32,uint8,uint64)": TypedContractEvent<
      TopicViewUpdatedEvent.InputTuple,
      TopicViewUpdatedEvent.OutputTuple,
      TopicViewUpdatedEvent.OutputObject
    >;
    TopicViewUpdated: TypedContractEvent<
      TopicViewUpdatedEvent.InputTuple,
      TopicViewUpdatedEvent.OutputTuple,
      TopicViewUpdatedEvent.OutputObject
    >;
  };
}